+++ Cargo.toml
@@ -17,5 +17,8 @@ serde_json = "1.0.133"
 ssh2 = { version = "0.9.4", features = ["vendored-openssl"] }
 tokio = { version = "1.42.0", features = ["io-std", "macros", "rt"] }
 
+[dev-dependencies]
+tempfile = "3.14.0"
+
 [profile.release]
 lto = "fat"
