+++ src/modules/template.rs
@@ -60,3 +60,82 @@ pub fn template(lua: &Lua, params: Table) -> mlua::Result<Table> {
 
     Ok(module)
 }
+
+// Tests
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use mlua::Lua;
+    use std::io::Write;
+    use tempfile::NamedTempFile;
+
+    #[test]
+    fn test_template_src_required() {
+        let lua = Lua::new();
+        let params = lua.create_table().unwrap();
+        let result = template(&lua, params);
+        assert!(result.is_err());
+        assert_eq!(
+            result.unwrap_err().to_string(),
+            "runtime error: 'src' parameter is required"
+        );
+    }
+
+    #[test]
+    fn test_template_dst_required() {
+        let lua = Lua::new();
+        let params = lua.create_table().unwrap();
+        params.set("src", "example.src").unwrap();
+        let result = template(&lua, params);
+        assert!(result.is_err());
+        assert_eq!(
+            result.unwrap_err().to_string(),
+            "runtime error: 'dst' parameter is required"
+        );
+    }
+
+    #[test]
+    fn test_template_vars_must_be_table() {
+        let lua = Lua::new();
+        let params = lua.create_table().unwrap();
+        params.set("src", "example.src").unwrap();
+        params.set("dst", "example.dst").unwrap();
+        params.set("vars", "not a table").unwrap();
+        let result = template(&lua, params);
+        assert!(result.is_err());
+        assert_eq!(
+            result.unwrap_err().to_string(),
+            "runtime error: 'vars' parameter must be a table"
+        );
+    }
+
+    #[test]
+    fn test_template_src_file_exists() {
+        let lua = Lua::new();
+        let params = lua.create_table().unwrap();
+        params.set("src", "non_existent_file.src").unwrap();
+        params.set("dst", "example.dst").unwrap();
+        let result = template(&lua, params);
+        assert!(result.is_err());
+        assert_eq!(
+            result.unwrap_err().to_string(),
+            "runtime error: Source template does not exist"
+        );
+    }
+
+    #[test]
+    fn test_template_success() {
+        let mut temp_file = NamedTempFile::new().unwrap();
+        writeln!(temp_file, "{{ name }} is {{ age }} years old").unwrap();
+        let lua = Lua::new();
+        let params = lua.create_table().unwrap();
+        params.set("src", temp_file.path().to_str().unwrap()).unwrap();
+        params.set("dst", "/remote/file").unwrap();
+        let vars = lua.create_table().unwrap();
+        vars.set("name", "John").unwrap();
+        vars.set("age", 30).unwrap();
+        params.set("vars", vars).unwrap();
+        let result = template(&lua, params);
+        assert!(result.is_ok());
+    }
+}
